socialnetwork/requirements.txt
asgiref==3.4.1
channels==3.0.4
Django==3.0.3
djangorestframework==3.13.1
factory==1.2
factory_boy==3.2.1
celery==5.2.3
redis==4.1.4
channels-redis==3.3.1
Pillow==9.0.1
